+++ libdieharder/countx.c
@@ -232,7 +232,7 @@ int main_countx( int argc, char **argv)
   free(data);
 
   time(&z);
-  printf("number of seconds: %6lu\n", (size_t)(z-a));
+  printf("number of seconds: %6lu\n", (unsigned long)(z-a));
 
   return 0;
