+++ libdieharder/rng_file_input.c
@@ -168,7 +168,9 @@ static unsigned long int file_input_get (void *vstate)
    state->rptr++;
    state->rtot++;
    if(verbose){
-	 fprintf(stdout,"# file_input() %lu: %lu/%lu -> %u\n", state->rtot, state->rptr,state->flen,(uint)iret);
+     fprintf(stdout,"# file_input() %lu: %lu/%lu -> %u\n",
+             (unsigned long)state->rtot, (unsigned long)state->rptr,
+             (unsigned long)state->flen,(uint)iret);
    }
 
    /*
@@ -276,7 +278,7 @@ static void file_input_set (void *vstate, unsigned long int s)
 	 state->rewind_cnt++;
 	 if(verbose == D_FILE_INPUT || verbose == D_ALL){
 	   fprintf(stderr,"# file_input(): Rewinding %s at rtot = %u\n", filename,(uint) state->rtot);
-	   fprintf(stderr,"# file_input(): Rewind count = %u, resetting rptr = %lu\n",state->rewind_cnt,state->rptr);
+	   fprintf(stderr,"# file_input(): Rewind count = %u, resetting rptr = %lu\n",state->rewind_cnt, (unsigned long)state->rptr);
 	 }
    } else {
 	 return;
@@ -328,7 +330,7 @@ static void file_input_set (void *vstate, unsigned long int s)
 	   cnt++;
 	   if(verbose){
 		 fprintf(stdout,"# file_input(): cnt = %d\n",cnt);
-		 fprintf(stdout,"# file_input(): state->flen set to %lu\n",state->flen);
+		 fprintf(stdout,"# file_input(): state->flen set to %lu\n",(unsigned long)state->flen);
 	   }
 	 }
 	 if(strncmp(splitbuf[0],"numbit",6) == 0){
